pyproject.toml
[project]
name = "flowchart-creator-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "html2image>=2.0.5",
    "ipykernel>=6.29.5",
    "langchain-groq>=0.2.3",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
    "streamlit-chat>=0.1.1",
    "streamlit>=1.41.1",
    "streamlit-extras>=0.5.0",
    "langchain-openai>=0.2.9",
    "langchain==0.3.7",
    "langchain-community>=0.3.7",
    "langgraph==0.2.34",
]
